+++ inc/clientinjection.class.php
@@ -41,6 +41,7 @@ use function Safe\json_decode;
 use function Safe\json_encode;
 use function Safe\readfile;
 use function Safe\unlink;
+use function Safe\unserialize;
 
 class PluginDatainjectionClientInjection
 {
@@ -217,7 +218,15 @@ class PluginDatainjectionClientInjection
 
         for ($i = $offset; $i < $end; $i++) {
             $injectionline = $i + $header_offset;
-            $result        = $engine->injectLine($lines[$i][0], $injectionline);
+            try {
+                $result = $engine->injectLine($lines[$i][0], $injectionline);
+            } catch (Throwable $e) {
+                ErrorHandler::logCaughtException($e);
+                $result = [
+                    'status' => PluginDatainjectionCommonInjectionLib::FAILED,
+                    'line'   => $injectionline,
+                ];
+            }
             $results[]     = $result;
 
             if ($result['status'] != PluginDatainjectionCommonInjectionLib::SUCCESS) {